# modules/schema/openapi.json
{
  "components": {
...
    "schemas": {
...
      "Omit_UserEntity.password_": {
        "$ref": "#/components/schemas/Pick_UserEntity.Exclude_keyofUserEntity.password__"
      },
      "UserState": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Omit_UserEntity.password_"
          },
          {
            "properties": {
              "__typename": {
                "type": "string",
                "enum": [
                  "User"
                ],
                "nullable": false
              }
            },
            "required": [
              "__typename"
            ],
            "type": "object"
          }
        ]
      },
      "LoginParams": {
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "UserParams": {
        "$ref": "#/components/schemas/Omit_UserEntity.id_"
      },
      "Partial_UserParams_": {
        "properties": {
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          }
        },
        "type": "object",
        "description": "Make all properties in T optional"
      },
...
      "ProjectParams": {
        "$ref": "#/components/schemas/Omit_ProjectEntity.id-or-contextId-or-userId-or-parentId-or-stub_"
      },
...
      "ServerState": {
        "properties": {
          "user": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ServerUserState"
              }
            ],
            "nullable": true
          },
          "problems": {
            "items": {
              "$ref": "#/components/schemas/Problem"
            },
            "type": "array"
          },
          "schemaVersion": {
            "type": "string"
          }
        },
        "required": [
          "user",
          "problems",
          "schemaVersion"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "Omit_ApiItemFilter.itemHolderId_": {
        "$ref": "#/components/schemas/Pick_ApiItemFilter.Exclude_keyofApiItemFilter.itemHolderId__"
      },
...
    },
    "securitySchemes": {}
  },
  "info": {
    "title": ""
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/page": {
      "get": {
        "operationId": "GetPageContent",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/login": {
      "post": {
        "operationId": "Login",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserState"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginParams"
              }
            }
          }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "operationId": "ListUsers",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserState"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": []
      },
      "put": {
        "operationId": "CreateUser",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserState"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserParams"
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "EditUser",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserState"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Partial_UserParams_"
                  },
                  {
                    "properties": {
                      "currentPassword": {
                        "type": "string"
                      },
                      "id": {
                        "type": "string"
                      }
                    },
                    "type": "object"
                  }
                ]
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "DeleteUser",
        "responses": {
          "204": {
            "description": "No content"
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "id": {
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/state": {
      "post": {
        "operationId": "GetState",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServerState"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "itemFilter": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/Omit_ApiItemFilter.itemHolderId_"
                      },
                      {
                        "items": {
                          "$ref": "#/components/schemas/Omit_ApiItemFilter.itemHolderId_"
                        },
                        "type": "array"
                      }
                    ]
                  }
                },
                "required": [
                  "itemFilter"
                ],
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/project": {
      "put": {
        "operationId": "CreateProject",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectState"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "params": {
                    "$ref": "#/components/schemas/ProjectParams"
                  },
                  "taskListId": {
                    "type": "string"
                  }
                },
                "required": [
                  "params",
                  "taskListId"
                ],
                "type": "object"
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "EditProject",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectState"
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "params": {
                    "$ref": "#/components/schemas/Partial_ProjectParams_"
                  },
                  "id": {
                    "type": "string"
                  }
                },
                "required": [
                  "params",
                  "id"
                ],
                "type": "object"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "DeleteProject",
        "responses": {
          "204": {
            "description": "No content"
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "id": {
                    "type": "string"
                  }
                },
                "required": [
                  "id"
                ],
                "type": "object"
              }
            }
          }
        }
      }
    },
    "/api/project/contents": {
      "post": {
        "operationId": "ListContents",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ProjectContents"
                    },
                    {
                      "$ref": "#/components/schemas/ContextContents"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "itemFilter": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/Omit_ApiItemFilter.itemHolderId_"
                      },
                      {
                        "items": {
                          "$ref": "#/components/schemas/Omit_ApiItemFilter.itemHolderId_"
                        },
                        "type": "array"
                      }
                    ]
                  },
                  "id": {
                    "type": "string"
                  }
                },
                "required": [
                  "id"
                ],
                "type": "object"
              }
            }
          }
        }
      }
    },
...
  }
}